requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
openpyxl>=3.1.0
lxml>=4.9.0
playwright>=1.45.0
playwright-stealth>=1.0.6
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Session==0.5.0
Werkzeug==3.0.0
mysql-connector-python>=9.0.0
seleniumbase>=4.0.0
google-generativeai>=0.3.0
pydantic>=2.6.0
html2text>=2024.2.26
google-genai>=1.0.0
